requirements.txt
langchain==0.1.11
openai
tiktoken
faiss-cpu
heyoo
Office365-REST-Python-Client
pylint
autopep8
black
azure-functions
langchain-openai
langchain-community
pytest
flask
azure-cosmos
azure-storage-blob
pandas
psycopg[binary]
pgvector
openpyxl